artwork_gallery_project/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bbcode==1.1.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dj-database-url==0.5.0
Django==4.0.3
django-environ==0.8.1
django-heroku==0.3.1
idna==3.3
mypy-extensions==0.4.3
nose==1.3.7
pathspec==0.9.0
platformdirs==2.5.1
psycopg2==2.9.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
whitenoise==6.0.0
